Abprob {DTMC}	R Documentation

Calculating probability when the absorbing state is known Calculate the conditional probability of transferring n steps from state i to state j, given never entering absorb state in this n steps. In real life, there may be known discrete Markov chains that never enter a state k1, k2... This function can help calculate the conditional probability of moving from state i to state j in n steps in this case. All we need to do is considering the states k1,k2... as absorbing state to put into this function and enter the transition matrix P and state i and state j. The function outputs the final probability. If absorbing state contain state i or j, the function will return the probability of having entered the absorbing states in these n steps.

Description

Calculating probability when the absorbing state is known Calculate the conditional probability of transferring n steps from state i to state j, given never entering absorb state in this n steps. In real life, there may be known discrete Markov chains that never enter a state k1, k2... This function can help calculate the conditional probability of moving from state i to state j in n steps in this case. All we need to do is considering the states k1,k2... as absorbing state to put into this function and enter the transition matrix P and state i and state j. The function outputs the final probability. If absorbing state contain state i or j, the function will return the probability of having entered the absorbing states in these n steps.

Usage


Abprob(P, Abstate, i, j, n = 1)

Arguments

P
transition matrix

Abstate
a vector contain all absorbing state colunm numbers

i
initial state column number

j
terminal state column number

n
Number of transfer steps. the default value is 1

Value

the conditional probability of transferring n steps from state i to state j, given never entering absorb state in this n steps

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
prob <- Abprob(P, Abstate = c(3), i = 1, j = 2, n = 5)


[Package DTMC version 0.9.1 Index]