CKequ {DTMC}	R Documentation

Champman-Kolmogorov Equations

Description

Calculate the transition probability from one state to another state.

Usage


CKequ(P, i, j, n)

Arguments

P
The transition probability matrix

i
The starting state

j
The ending state

n
The number of steps

Details

The input matrix should be a transition probability matrix of discrete time Markov chain, the starting and ending states should be numerical values within the range of row(column) indices, and the step is the number of periods during the process.

Value

The probability from state i to state j in n steps

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
prob <- CKequ(P = P, i = 1, j = 3, n = 2)


[Package DTMC version 0.9.1 Index]