DTMC {DTMC}	R Documentation

DTMC: A package for finite discrete markov chain to solve stochastic problems.

Description

Introduction: The main goals of DTMC package include:

Judge whether a stochastic process is a discrete time Markov chain.

Classification of states in the Markov chain, like transient or recurrent, accessible or communicate.

Calculation of transition probabilities, long-run proportions, future expectations, and so on.

Graphical visualization of the Markov chain.

Details

Formation of Probability Matrix Users can input a transition probability matrix and judge whether it is a valid DTMC by using is.DTMC(). Users can also choose to randomly generate a transition probability matrix which satisfies the requirements of a valid DTMC by using runif_DTMC().

Classification of States It is important to classify different states in a DTMC. To give a more direct response to users, the package provide some graphic visualizations to represent the Markov chain. If it is possible to change from state i to j, j is said to be accessible from i, and this can be checked by is.accessible(). If two states are accessible from each other, they are said to communicate, and this can be checked by is.communicate(). If all states in a Markov can communicate with other states, these states are said to form a class, and this can be checked by is.class(). If users want to have a graphical view of the relationship between different states, they can use plot_graph() to visualize the result.

Statistical Calculation Transition probability from one state to another in n steps can be calculated by CKequ(). Calculate probabilities especially for absorbing states by using Abprob(). Since states can be classified as recurrent and transient. For transient states, users can get the average number of steps from transient i to state j by TranTime().

Long-run Behavior The DTMC is said to reach a stabilized condition if after infinite many steps, probability in each state follows certain distribution. If long-run proportions exist, for given probability mass function, the long-run expected value of the DTMC can be calculated by expect_DTMC() and a value will be returned. To compute the long-run proportions of all states in the Markov chain, users can use LongRunP() and a vector of probabilities will be returned. Suppose the process from state i to state j is the forward Markov chain, then the process from state j to state i is the backward chain. To compute the backward chain, users can use RevChain() and the reversed probability matrix will be returned. If the forward chain and backward chain satisfy certain properties under the stationary distribution, whether the process is time reversible can be checked by using isTReversible().

[Package DTMC version 0.9.1 Index]