LongRunP {DTMC}	R Documentation

Long-run proportion

Description

Calculate the long-run probability of each state if the Markov chain is irreducible and recurrent.

Usage


LongRunP(P, sign = F)

Arguments

P
The transition probability matrix

sign
If the transition matrix being time reversible has been known, sign is TRUE; otherwise sign is FALSE

Details

To calculate the long-run proportion, since there are finite states, the Markov chain must be irreducible and recurrent. If those conditions are satisfied, the function will return a vector of stationary probabilities for each state, otherwise there will be no result.

Value

A vector of long-run probabilities of each state or NA if probabilities do not exist

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
LRp <- LongRunP(P = P, sign = TRUE)


[Package DTMC version 0.9.1 Index]