RevChain {DTMC}	R Documentation

Get the reversed Markov chain

Description

Derive the reversed chain of the original Markov chain

Usage


RevChain(P)

Arguments

P
The transition probability matrix

Details

The input matrix should be a transition probability matrix. If the long-run proportion of the matrix exists and satisfies certain properties, the Markov chain has a reversed chain.

Value

The reversed transition probability matrix or NA if does not exist

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
revP <- RevChain(P)


[Package DTMC version 0.9.1 Index]