TranTime {DTMC}	R Documentation

Mean time spent in transient states

Description

Description: Calculate the average number of periods spent in transient states

Usage


TranTime(Pt, i, j)

Arguments

Pt
The transient transition probability matrix

i
The starting transient state i

j
The ending transient state j

Details

Details: For transient states in a Markov chain, limiting probabilities do not exist, but the average time from transient state i to transient state j can be calculated. Given the transient probability matrix, transient states i and j, the expected amount of time from i to j will be returned.

Value

The mean time spent from transient state i to j or NA if does not exist

Examples


Pt <- matrix(c(0, 0.4,   0,   0,   0,   0,
             0.6,   0, 0.4,   0,   0,   0,
               0, 0.6,   0, 0.4,   0,   0,
               0,   0, 0.6,   0, 0.4,   0,
               0,   0,   0, 0.6,   0, 0.4,
               0,   0,   0,   0,   0.6, 0), nrow = 6, byrow = TRUE)
tranT <- TranTime(Pt = Pt, i = 3, j = 5)

[Package DTMC version 0.9.1 Index]