class_info {DTMC}	R Documentation

Get class information

Description

Get class information of a whole discrete time Markov chains

Usage


class_info(P)

Arguments

P
the whole transition matrix P

Details

The user needs to input the whole transition matrix, and this function can generate all the classification information of the discrete time Markov chain, including the column number of the state contained by each class after classification, the corresponding block transition matrix of each class, the classification of each class and a vector contain the group number of each state in order. This information is contained in a list.

Value

return a list

list$states contains the column number of the state contained by each class after classification

list$prob_martix contains the corresponding block transition matrix of each class

list$class contains the classification of each class

list$group a vector contain the group number of each state in order

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0, 1), byrow = TRUE, nrow = 3)
class_info(P)


[Package DTMC version 0.9.1 Index]