classify {DTMC}	R Documentation

Classification of a class of states

Description

Classify a class of states into recurrent or transient

Usage


classify(P)

Arguments

P
transition matrix P or a block matrix P of the whole transition matrix whose state should be in the same class

Details

State is said to be recurrent if it have probability to return itself in the future. Otherwise, it is transient. The function can help users get class type of a given class. The input P is a transition matrix or a block matrix P of the whole transition matrix whose state should be in the same class. Otherwise, the function returns NA

Value

if all states are in the same class, return a string ("recurrent" or "transient"). Otherwise, renturn NA.

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
print(classify(P))


[Package DTMC version 0.9.1 Index]