expect_DTMC {DTMC}	R Documentation

Compute the expectation.

Description

expect_DTMC computes the expectation of markov chain with the given state real value. It also allows the user to input a function with respect to state true value.

Usage


expect_DTMC(P, state_value = 1:nrow(P), FUN = function(x) x)

Arguments

P
a transition probability matrix

state_value
the real value of each state. Must be finite vector, same size with nrow(P). The default value is columns numbers of states

FUN
a function object whose input is the real value of state. The default value is function(x) x

Details

If the expectation is finite, the expect value E(g(x)) will be returned, where x is the input real value of states and g(x) is the input FUN. While Function is unbounded, the expectation will be shown as "does not exist" rather than returned as Inf.

Value

The expectation value of the markov chain with given function FUN.

Examples


P = matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
state_value = c(2, 4, 6)
FUN = (function(x) x^2)
expect_DTMC(P, state_value, FUN)


[Package DTMC version 0.9.1 Index]