get_graph {DTMC}	R Documentation

Get the corresponding graph

Description

Get the corresponding graph of this discrete Markov chain.

Usage


get_graph(P)

Arguments

P
transition matrix P

Details

Discrete Markov chains can be represented by weighted graphs of graph theory. That is, states are vertices of graphs. If a state i can reach a state j through a one-step transfer (j can be equal to i), there will be a directed line from state i to state j. The weight is the transition probability from state i to state j. This function converts the discrete Markov chain through the igraph package into a graph object in the igraph package. The user can use this object to process it using a function in the igraph package.

Value

a graph object of igraph

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
g <- get_graph(P)
print(g)


[Package DTMC version 0.9.1 Index]