is.DTMC {DTMC}	R Documentation

Check whether the matrix is valid

Description

is.DTMC checks whether the input matrix is valid. The DTMC package only solves finite and non-negative matrix, with each row sum to be 1.

Usage


is.DTMC(P)

Arguments

P
a transition probability matrix.

Details

The check order is as following:

Matrix must be a square matrix.

Matrix must have each row sum 1.

Matrix's elements must all be larger or equal to 0 and smaller or equal to 1.

Value

Whether the input transition matrix is valid or not.

Examples


P = matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
is.DTMC(P)


[Package DTMC version 0.9.1 Index]