is.accessible {DTMC}	R Documentation

Check the accessibility of the state

Description

Check whether state j is accessible from state i.

Usage


is.accessible(P, i, j)

Arguments

P
transition matrix P or a block matrix P of the whole transition matrix

i
initial state column number

j
terminal state column number

Details

That state j is accessible to state i means that there is a step from state i that is possible to move to state j. The state i should not equal to state j. If you want to verify that a state can return to itself, call class_info() to verify that the state is transient.

Value

return NA if i is equivalent with j, and return True if state j is accessible to state i. Otherwise, return False.

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
if (is.accessible(P, 1, 3)) {
   print(P[c(1, 3), c(1, 3)])
}


[Package DTMC version 0.9.1 Index]