is.class {DTMC}	R Documentation

Class Check

Description

Check if some states are in the same class

Usage


is.class(P, states = NA)

Arguments

P
transition matrix P or a block matrix P of the whole transition matrix

states
a vector contain all state column number. The default value is NA. If states is NA, this is equivalent to entering all the state column numbers

Details

State in the same class means that they can all communicate with each other. You can input transition matrix P or block matrix P of the whole transition matrix, and states is a vector contain all state column number of that you want to check. If you do not enter the parameter "states", it will enter NA, meaning that the function will check whether all states in P are in the same class. It is worth mentioning that if P is the entire transition matrix and states is NA, then this function checks whether the DTMC is an irreducible.

Value

return True if all state in states are in the same class. Otherwise, return False.

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
if (is.class(P)) {
   print("The DTMC is irreducible.")
}


[Package DTMC version 0.9.1 Index]