isTReversible {DTMC}	R Documentation

Judge whether the Markov chain is time reversible

Description

Returns whether the Markov matrix present in the argument is time reversible.

Usage


isTReversible(P)

Arguments

P
The transition probability matrix

Details

The input transition matrix should be in the form of n by n. If the matrix satisfies certain property, the markov chain can be classified as time reversible.

Value

A boolean value indicating whether the Markov chain is time reversible

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
judge <- isTReversible(P)


[Package DTMC version 0.9.1 Index]