plot_graph {DTMC}	R Documentation

Plot a graph

Description

Plot a graph of the DTMC

Usage


plot_graph(P, ...)

Arguments

P
transition matrix P

...
some setup of plot and the setup of plot in the igraph() package

Details

This function calls get_graph(P) to get the graph associated with P, and then plots it with state classes reasonably. States in the same class are circled in the same color.

Value

a list contains that class information obtain from class_info(), and plot a graph

Examples


P <- matrix(c(0.3, 0.3, 0.4, 0.1, 0.2, 0.7, 0, 0.2, 0.8), byrow = TRUE, nrow = 3)
plot_graph(P)


[Package DTMC version 0.9.1 Index]