runif_DTMC {DTMC}	R Documentation

Create a transition matrix randomly.

Description

runif_DTMC creates a valid probability transition matrix randomly.

Usage


runif_DTMC(nrow = 3)

Arguments

nrow
the desired number of rows (or columns) in the square matrix. Must be finite. The default value is 3.

Details

The matrix randomly generated must be valid probability transition matrix. It must be finite and non-negative, with each of the row sum being 1.

Value

The random probability transition matrix generated.

Examples


(P <- runif_DTMC(nrow = 5))
is.DTMC(P)
rowSums(P)
colSums(P)


[Package DTMC version 0.9.1 Index]